Excluded 'Politics==8'
The question is supposed to be answered on a scale from 1-7, but 57 people in the Japan survey answered 8. It's hard to know why this happened so this version excludes them. I could see a case for simply re-scaling all of the Japan Politics responses, and that could still be done with the previous datasets.